refactor(settings): move scan variables to Compliance policies
Variable values define WHAT compliant means for the organization
(banner text, password policy, NTP pool) - policy content - while the
Scanning page owns WHEN scans run. ScanVariablesCard extracts to
components/settings/; PoliciesPage graduates from the stub file
(live Scan variables section + per-section pending banners on the
framework-lenses and exception-workflow stubs).

Specs: frontend-settings-scan-config AC-06 re-homed (unreleased on
this branch); frontend-settings v1.2.0 (AC-15: five stubs, PoliciesPage
graduated).
